--- src/handler/discord.js/commands/index.ts
@@ -1,18 +1,0 @@
-export { ReminderSetCommandHandler } from "@/src/handler/discord.js/commands/ReminderSetCommandHandler";
-export { ReminderListCommandHandler } from "@/src/handler/discord.js/commands/ReminderSetCommandHandler";
-export { ReminderDeleteCommandHandler } from "@/src/handler/discord.js/commands/ReminderSetCommandHandler";
-export { ReviewGachaCommandHandler } from "@/src/handler/discord.js/commands/ReviewGachaCommandHandler";
-export { ReviewListCommandHandler } from "@/src/handler/discord.js/commands/ReviewListCommandHandler";
-export { TranslateCommandHandler } from "@/src/handler/discord.js/commands/TranslateCommandHandlers";
-export { ChoiceCommandHandler } from "@/src/handler/discord.js/commands/ChoiceCommandHandler";
-export { DiceCommandHandler } from "@/src/handler/discord.js/commands/DiceCommandHandler";
-export { HelpCommandHandler } from "@/src/handler/discord.js/commands/HelpCommandHandler";
-export { MinecraftStartCommandHandler } from "@/src/handler/discord.js/commands/MinecraftStartCommandHandler";
-export { MinecraftStopCommandHandler } from "@/src/handler/discord.js/commands/MinecraftStartCommandHandler";
-export { ParrotCommandHandler } from "@/src/handler/discord.js/commands/ParrotCommandHandler";
-export { PointCheckCommandHandler } from "@/src/handler/discord.js/commands/PointCheckCommandHandler";
-export { PointExchangeCommandHandler } from "@/src/handler/discord.js/commands/PointExchangeCommandHandler";
-export { PointDrawCommandHandler } from "@/src/handler/discord.js/commands/PointDrawCommandHandler";
-export { PointItemCommandHandler } from "@/src/handler/discord.js/commands/PointItemCommandHandler";
-export { TalkCommandHandler } from "@/src/handler/discord.js/commands/TalkCommandHandler";
-export { WaiwaiCommandHandler } from "@/src/handler/discord.js/commands/WaiwaiCommandHandler";
