+++ src/handlers/discord.js/commands/ChoiceCommandHandler.ts
@@ -1,23 +1,27 @@
 import { LogicTypes } from "@/src/entities/constants/DIContainerTypes";
-import { HelpCategory } from "@/src/entities/vo/HelpCategory";
-import type { SlashCommandHandler } from "@/src/handler/discord.js/commands/SlashCommandHandler";
+import { ChoiceContent } from "@/src/entities/vo/ChoiceContent";
+import type { SlashCommandHandler } from "@/src/handlers/discord.js/commands/SlashCommandHandler";
 import type { IUtilityLogic } from "@/src/logics/Interfaces/logics/IUtilityLogic";
 import type { CacheType, ChatInputCommandInteraction } from "discord.js";
 import { inject, injectable } from "inversify";
 
 @injectable()
-export class HelpCommandHandler implements SlashCommandHandler {
+export class ChoiceCommandHandler implements SlashCommandHandler {
 	@inject(LogicTypes.UtilityLogic)
 	private utilLogic!: IUtilityLogic;
 	isHandle(commandName: string): boolean {
-		return commandName === "help";
+		return commandName === "choice";
 	}
+
 	async handle(
 		interaction: ChatInputCommandInteraction<CacheType>,
 	): Promise<void> {
 		await interaction.reply(
-			await this.utilLogic.help(
-				new HelpCategory(interaction.options?.getString("category", true)),
+			await this.utilLogic.choice(
+				interaction.options
+					.getString("items", true)
+					.split(" ")
+					.map((r) => new ChoiceContent(r)),
 			),
 		);
 	}
