+++ src/handlers/discord.js/commands/DiceCommandHandler.ts
@@ -1,6 +1,6 @@
 import { LogicTypes } from "@/src/entities/constants/DIContainerTypes";
 import { DiceSides } from "@/src/entities/vo/DiceSides";
-import type { SlashCommandHandler } from "@/src/handler/discord.js/commands/SlashCommandHandler";
+import type { SlashCommandHandler } from "@/src/handlers/discord.js/commands/SlashCommandHandler";
 import type { IUtilityLogic } from "@/src/logics/Interfaces/logics/IUtilityLogic";
 import type { CacheType, ChatInputCommandInteraction } from "discord.js";
 import { inject, injectable } from "inversify";
