+++ src/handlers/discord.js/commands/HelpCommandHandler.ts
@@ -1,24 +1,23 @@
 import { LogicTypes } from "@/src/entities/constants/DIContainerTypes";
-import { ParrotMessage } from "@/src/entities/vo/ParrotMessage";
-import type { SlashCommandHandler } from "@/src/handler/discord.js/commands/SlashCommandHandler";
+import { HelpCategory } from "@/src/entities/vo/HelpCategory";
+import type { SlashCommandHandler } from "@/src/handlers/discord.js/commands/SlashCommandHandler";
 import type { IUtilityLogic } from "@/src/logics/Interfaces/logics/IUtilityLogic";
 import type { CacheType, ChatInputCommandInteraction } from "discord.js";
 import { inject, injectable } from "inversify";
 
 @injectable()
-export class ParrotCommandHandler implements SlashCommandHandler {
+export class HelpCommandHandler implements SlashCommandHandler {
 	@inject(LogicTypes.UtilityLogic)
 	private utilLogic!: IUtilityLogic;
 	isHandle(commandName: string): boolean {
-		return commandName === "parrot";
+		return commandName === "help";
 	}
-
 	async handle(
 		interaction: ChatInputCommandInteraction<CacheType>,
 	): Promise<void> {
 		await interaction.reply(
-			await this.utilLogic.parrot(
-				new ParrotMessage(interaction.options?.getString("message", true)),
+			await this.utilLogic.help(
+				new HelpCategory(interaction.options?.getString("category", true)),
 			),
 		);
 	}
