+++ src/handlers/discord.js/commands/ParrotCommandHandler.ts
@@ -1,27 +1,24 @@
 import { LogicTypes } from "@/src/entities/constants/DIContainerTypes";
-import { ChoiceContent } from "@/src/entities/vo/ChoiceContent";
-import type { SlashCommandHandler } from "@/src/handler/discord.js/commands/SlashCommandHandler";
+import { ParrotMessage } from "@/src/entities/vo/ParrotMessage";
+import type { SlashCommandHandler } from "@/src/handlers/discord.js/commands/SlashCommandHandler";
 import type { IUtilityLogic } from "@/src/logics/Interfaces/logics/IUtilityLogic";
 import type { CacheType, ChatInputCommandInteraction } from "discord.js";
 import { inject, injectable } from "inversify";
 
 @injectable()
-export class ChoiceCommandHandler implements SlashCommandHandler {
+export class ParrotCommandHandler implements SlashCommandHandler {
 	@inject(LogicTypes.UtilityLogic)
 	private utilLogic!: IUtilityLogic;
 	isHandle(commandName: string): boolean {
-		return commandName === "choice";
+		return commandName === "parrot";
 	}
 
 	async handle(
 		interaction: ChatInputCommandInteraction<CacheType>,
 	): Promise<void> {
 		await interaction.reply(
-			await this.utilLogic.choice(
-				interaction.options
-					.getString("items", true)
-					.split(" ")
-					.map((r) => new ChoiceContent(r)),
+			await this.utilLogic.parrot(
+				new ParrotMessage(interaction.options?.getString("message", true)),
 			),
 		);
 	}
