+++ src/routes/discordjs/events/ReactionRouter.ts
@@ -2,7 +2,7 @@ import { HandlerTypes } from "@/src/entities/constants/DIContainerTypes";
 import type {
 	DiscordEventHandler,
 	ReactionInteraction,
-} from "@/src/handler/discord.js/events/DiscordEventHandler";
+} from "@/src/handlers/discord.js/events/DiscordEventHandler";
 import type { DiscordEventRouter } from "@/src/routes/discordjs/events/DiscordEventRouter";
 import type { Client } from "discord.js";
 import { injectable, multiInject } from "inversify";
